refactor: modal 상태 hook 분리 및 modal 레이아웃 작업
