+++ src/app/test/page.tsx
@@ -1,25 +1,17 @@
 'use client';
 
-import { useState } from 'react';
+import useModalState from '@/hooks/useModalState';
 
 import Button from '@/components/button/Button';
 import Modal from '@/components/modal/Modal';
 
 const TestPage = () => {
-  const [modalOpen, setModalOpen] = useState(false);
-
-  const handleModalOpen = () => {
-    setModalOpen(true);
-  };
-
-  const handleModalClose = () => {
-    setModalOpen(false);
-  };
+  const { modalOpen, handleModalOpen, handleModalClose } = useModalState();
 
   return (
     <div>
       <Button onClick={handleModalOpen}>modal open</Button>
-      <Modal active={modalOpen} onClose={handleModalClose}>
+      <Modal active={modalOpen} onClose={handleModalClose} closeButton>
         <div>모달 열려따</div>
       </Modal>
     </div>
